IdP/requirements.txt
# IdP Project Requirements
Django==5.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
django-cors-headers==4.4.0
python-decouple==3.8
psycopg2-binary==2.9.9
requests==2.31.0
cryptography==41.0.8